Fix oversight when matching unquoted `href=values`
Whitespace wasn't being checked before and after the equals-sign, and an
unquoted value matched directly before a tag's closing ">" would consume
the rest of the tag. Noticed when testing the regexp on this:

   <A HREF = http://url></a>

Clearer and annotated version of `cleanHrefs()` added as a commented sed
script. Should be clearer for my future-self, hopefully.