Force ANSI escape to precede newline in APM output
This causes problems when immediately followed by an `::endgroup::` line
(GitHub Actions expects commands to appear immediately after a newline).
A kludgey attempt to solve this was attempted by 2b47c39, which caused a
blank line to appear at the end of every other group. It's been reverted
in favour of a more effective solution.

In addition, don't print error messages to STDERR since it's interleaved
with other messages, making it harder to locate the source of the error.