Switch to `test` directory before running CI tasks
This makes the workflow example mentioned in `README.md` a bit confusing
to anybody expecting a fully-reusable workflow template. We can axe this
step if we add support for specifying an arbitrary root directory via an
environment variable.

For now, readers will just have to deal with a documented inconsistency.